Championship view, model and menu action. Reaftoring and __init__file creation.
