+++ domain/championship/models.py
@@ -1,0 +1,8 @@
+from utils.graph import ReverseBinaryGraph
+
+__author__ = 'msoum'
+
+
+class ChampionshipModel:
+    def __init__(self, team_list):
+        self.graph = ReverseBinaryGraph(team_list)
