--- domain/contest_manager.py
@@ -1,6 +1,0 @@
-__author__ = 'Mathieu'
-
-
-class FourMatchesContestManager:
-    def __init__(self):
-        pass
