+++ domain/team/models.py
@@ -1,0 +1,86 @@
+from PySide.QtCore import QAbstractItemModel, QModelIndex, Qt
+from domain.data_structures import Team
+
+__author__ = 'msoum'
+
+
+class TeamModel(QAbstractItemModel):
+    def __init__(self):
+        super(TeamModel, self).__init__()
+
+        self.team_list = [
+            Team('A', 'A', 1),
+            Team('B', 'B', 2),
+            Team('C', 'C', 3),
+            Team('D', 'D', 4),
+            Team('E', 'E', 5),
+            Team('F', 'F', 6),
+            Team('G', 'G', 7),
+            Team('H', 'H', 8),
+            Team('I', 'I', 9),
+            Team('J', 'J', 10),
+            Team('K', 'K', 11),
+        ]
+
+    def columnCount(self, parent=QModelIndex):
+        return 2
+
+    def rowCount(self, parent=QModelIndex):
+        return len(self.team_list)
+
+    def data(self, index, role=Qt.DisplayRole):
+        if role == Qt.DisplayRole:
+            if index.column() == 0:
+                return self.team_list[index.row()].name
+            if index.column() == 1:
+                return self.team_list[index.row()].club
+        elif role == Qt.TextAlignmentRole:
+            return Qt.AlignCenter
+        else:
+            return None
+
+    def index(self, row, column, parent=QModelIndex):
+        return self.createIndex(row, column)
+
+    def parent(self, child):
+        return QModelIndex()
+
+    def headerData(self, section, orientation, role):
+        if orientation == Qt.Vertical:
+            return super(TeamModel, self).headerData(section, orientation, role)
+        elif orientation == Qt.Horizontal and role == Qt.DisplayRole:
+            if section == 0:
+                return 'Team name'
+            else:
+                return 'Club name'
+        else:
+            return None
+
+    def add_team(self, name, club):
+        self.beginInsertRows(QModelIndex(), len(self.team_list), len(self.team_list))
+        team = Team(name, club, len(self.team_list))
+        self.team_list.append(team)
+        self.endInsertRows()
+
+    def get_team(self, row):
+        return self.team_list[row]
+
+    def update_team(self, team, new_name, new_club):
+        self.beginResetModel()
+        team_idx = self.team_list.index(team)
+        self.team_list[team_idx] = Team(new_name, new_club, team.number)
+        self.endResetModel()
+
+    def delete_team(self, team):
+        team_idx = self.team_list.index(team)
+        self.beginRemoveRows(QModelIndex(), team_idx, team_idx)
+        self.team_list.pop(team_idx)
+        self.endRemoveRows()
+
+    def get_team_from_name(self, name):
+        if name == "None":
+            return Team()
+        if name == "###":
+            return Team("###", "###")
+        else:
+            return [it for it in self.team_list if it.name == name][0]
