+++ views/main_window.py
@@ -1,6 +1,7 @@
 from PySide.QtGui import QAction, QMainWindow
-from domain.models import team_model
-from views.widgets import RegistrationWidget, ContestWidget
+
+from domain.four_matches.models import team_model
+from views.widgets import RegistrationWidget, ContestWidget, ChampionshipWidget
 
 __author__ = 'Mathieu'
 
@@ -13,8 +14,10 @@ class MainWindow(QMainWindow):
 
         # Action menu
         self.action_menu = menu_bar.addMenu('A&ction')
-        self.generate_match_action = QAction('&Generate matches', self)
-        self.action_menu.addAction(self.generate_match_action)
+        self.generate_contest_action = QAction('Generate &contest', self)
+        self.generate_championship_action = QAction('Generate c&hampionship', self)
+        self.action_menu.addAction(self.generate_contest_action)
+        self.action_menu.addAction(self.generate_championship_action)
 
         # Team menu
         self.team_menu = menu_bar.addMenu('&Team')
@@ -37,7 +40,8 @@ class MainWindow(QMainWindow):
         self.setCentralWidget(central_widget)
 
         # Connect action to slots
-        self.generate_match_action.triggered.connect(self.start_contest)
+        self.generate_contest_action.triggered.connect(self.start_contest)
+        self.generate_championship_action.triggered.connect(self.start_championship)
         self.team_add_action.triggered.connect(central_widget.add_new_team_slot)
         self.team_edt_action.triggered.connect(central_widget.edt_team_slot)
         self.team_del_action.triggered.connect(central_widget.del_team_slot)
@@ -47,11 +51,21 @@ class MainWindow(QMainWindow):
         self.resize(500, 400)
 
     def start_contest(self):
-        self.generate_match_action.setEnabled(False)
+        self.generate_contest_action.setEnabled(False)
+        self.generate_championship_action.setEnabled(False)
         self.team_menu.setEnabled(False)
         self.status_action.setEnabled(True)
         contest_widget = ContestWidget()
         self.setCentralWidget(contest_widget)
         self.status_action.triggered.connect(contest_widget.show_status_view)
+        self.resize(900, 700)
+        self.showMaximized()
+
+    def start_championship(self):
+        self.generate_contest_action.setEnabled(False)
+        self.generate_championship_action.setEnabled(False)
+        self.team_menu.setEnabled(False)
+        championship_widget = ChampionshipWidget()
+        self.setCentralWidget(championship_widget)
         self.resize(900, 700)
         self.showMaximized()
